refactor wave report input router
